lang: A wrapper around boost assert.
This adds optionally (and by default) non-fatal assert macros, which more or less transparently wrap boost assert. In addition, SC_COND_ASSERT is an assert that (a) executes in both debug and release, and (b) results in the boolean value of the assert expression. This way, it can be used in cases where you want to assert in debug, and protect against an unexpected state in release. The most common form of this is:
if (SC_COND_ASSERT(objectPtr != NULL)) {
    *objectPtr;
}