lang: Refactor bytecode semantics
This removes the C-style pointer-heavy bytecode operations, and replaces them with more modern C++ patterns. Bytecode arrays are stored as deque's (a good match for our push-often iterate-rarely scenario), with an std allocator wrapped around the compile pool. ByteCodeBase objects are accessed via a unique_ptr (ByteCode), so all ownership transfers are now represented by a std::move operation. Library compilation performance is unchanged - in theory, it should be slightly faster as there are few copies being made, but in practice the difference is miniscule.