lang: Add source position table for PyrBlocks
This tracks source code positions during the compilation process, and accumulates those positions in PyrBlock::debugTable (visible in sclang as FunctionDef:debugTable). Every time we are preparing to enter a PyrParseNode::compile method, the source position of that PyrParseNode is updated globally. Any bytecodes that are compiled as a part of that ::compile call will be marked as having that source position. Source positions are collected with bytecodes, and attached to the PyrBlock in installByteCodes. Note that there is often significant repetition in the debug table, as many bytecodes might occur at the same line / char offset. More efficient storage of the debug table can be dealt with as an implementation detail for future iterations.